feat(messages): open chat in new page
